backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
google-genai>=1.70.0
supabase>=2.11.0
requests==2.32.3
beautifulsoup4==4.12.3
langchain==0.3.14
langchain-text-splitters==0.3.5
